Add files via upload
Simple and easily customizable script.  Simply run script and a folder will be created. Results will be uploaded to the folder in an already formatted xls file. for each year.

Just ran it once myself today to show what the files will look like in the folder. You may delete this folder and run the script again,